Fix warnings
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@100627 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/flatfile/utilfun.cpp
@@ -1043,9 +1043,9 @@ int SrchKeyword(const CTempString& ptr, const vector<string>& keywordList)
 {
     SIZE_TYPE keywordCount = keywordList.size();
 
-    for (int i = 0; i < keywordCount; ++i) {
+    for (unsigned i = 0; i < keywordCount; ++i) {
         if (NStr::StartsWith(ptr, keywordList[i])) {
-            return i;
+            return (int)i;
         }
     }
     return ParFlat_UNKW;
